PMS Support Reference
:Release Notes:
Change repo form pms-support-rpi to this
Implements states specific to Raspberry Pi
Adds state machine for reference as rpi.fsm

:Detailed Notes:
Moving code from pms-support-rpi to pms-support-reference
under webos-pro
For other devices it will be named as pms-support-<device>
This contains the reference specific states and its
implementation
All the supported states are defined in rpi.fsm XML file

Supported States:
1. ActiveState
2. ActiveStandbyState
3. RebootState
4. PowerOffState

Transitioning between states takes place depending
upon the event to be processed
It is created as a library and is plug-able to
power-manager-service.
It depends on fsm library and statemanager, part of libpmscore.

For ActiveState --> PowerOffState, first state machine moved to
intermediate state TransitionPowerOff to ask for voting, later
based on responses either moves to PowerOffState or
ActiveState (If any clients blocks it).

:Testing Performed:
Build successful. Test cases mentioned in [PLAT-43916].
Please refer the spreadsheet attached with it.

:QA Notes:

:Issues Addressed:
[PLAT-43916] Integrate PMS into image
[PLAT-41885] Convert the configuration file based on
             available power states for RPi
[PLAT-31877] Implement State transition for Poweroff

Change-Id: I3903546970079cbe31818fff166a1cf9e24495f3